JsonHelper is moved to a more generic package.
