+++ src/main/groovy/betsy/bpmn/engines/JsonHelper.java
@@ -1,4 +1,4 @@
-package betsy.bpmn.engines.camunda;
+package betsy.bpmn.engines;
 
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.JsonNode;
