+++ src/main/groovy/betsy/bpmn/engines/camunda/CamundaTester.java
@@ -1,9 +1,6 @@
 package betsy.bpmn.engines.camunda;
 
-import betsy.bpmn.engines.BPMNEnginesUtil;
-import betsy.bpmn.engines.BPMNTester;
-import betsy.bpmn.engines.LogFileAnalyzer;
-import betsy.bpmn.engines.OverlappingTimestampChecker;
+import betsy.bpmn.engines.*;
 import betsy.bpmn.model.BPMNAssertions;
 import betsy.bpmn.model.BPMNTestCase;
 import betsy.bpmn.model.BPMNTestVariable;
@@ -13,7 +10,6 @@ import org.apache.log4j.Logger;
 import org.json.JSONObject;
 
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.*;
 
 public class CamundaTester {
